+++ src/main/java/acorn/omakase/service/user/EmailService.java
@@ -1,4 +1,4 @@
-package acorn.omakase.service;
+package acorn.omakase.service.user;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.mail.javamail.JavaMailSender;
@@ -47,7 +47,7 @@ public class EmailService {
     private MimeMessage createEmailForm(String email) throws MessagingException, UnsupportedEncodingException {
 
         createCode(); //인증 코드 생성
-        String setFrom = "dj5668@naver.com"; //email-config에 설정한 자신의 이메일 주소(보내는 사람)
+        String setFrom = "pjacorn0513@naver.com"; //email-config에 설정한 자신의 이메일 주소(보내는 사람)
         String toEmail = email; //받는 사람
         String title = "CODEBOX 회원가입 인증 번호"; //제목
