+++ src/main/resources/templates/mail.html
@@ -4,7 +4,7 @@
 <body>
 <div style="margin:100px;">
     <h1> 안녕하세요.</h1>
-    <h1> 알고리즘 관리 서비스 CODEBOX 입니다.</h1>
+    <h1> 상권 분석 서비스 omakase 입니다.</h1>
     <br>
     <p> 아래 코드를 회원가입 창으로 돌아가 입력해주세요.</p>
     <br>
